# backend/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "VinFast API Documentation",
    "version": "1.0.0",
    "description": "\n        ## 🚗 VinFast API Documentation\n\n        API documentation cho website cửa hàng ô tô VinFast - Hệ thống quản lý bán hàng và dịch vụ ô tô điện.\n\n        ### 🔐 Authentication\n        API sử dụng JWT (JSON Web Token) để xác thực. Để sử dụng các endpoint được bảo vệ:\n        1. Đăng nhập qua `/api/auth/login` để lấy token\n        2. Thêm token vào header: `Authorization: Bearer <token>`\n\n        ### 📱 Features\n        - **Authentication & Authorization**: Đăng ký, đăng nhập, quản lý người dùng\n        - **Car Management**: Quản lý xe, tìm kiếm, lọc theo loại\n        - **Order Management**: Đặt hàng, theo dõi đơn hàng, thanh toán\n        - **Test Drive**: Đăng ký lái thử, quản lý lịch hẹn\n        - **Contact**: Hệ thống liên hệ tư vấn\n        - **News**: Quản lý tin tức, bài viết\n        - **User Management**: Quản lý người dùng (Admin)\n\n        ### 🌐 Base URL\n        - Development: `http://localhost:8000/api`\n        - Production: `https://api.vinfast.com/api`\n      ",
    "contact": {
      "name": "VinFast Development Team",
      "email": "dev@vinfast.com",
      "url": "https://vinfast.vn"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token để xác thực. Format: Bearer <token>"
      }
    },
    "schemas": {
...
      "Order": {
        "type": "object",
        "required": [
          "items",
          "customerInfo",
          "payment"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID đơn hàng"
          },
          "orderNumber": {
            "type": "string",
            "description": "Mã đơn hàng",
            "example": "VF20241201001"
          },
          "user": {
            "type": "string",
            "description": "ID người dùng"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "car": {
                  "type": "string"
                },
                "quantity": {
                  "type": "number"
                },
                "selectedColor": {
                  "type": "string"
                },
                "depositAmount": {
                  "type": "number"
                },
                "unitPrice": {
                  "type": "number"
                }
              }
            }
          },
          "customerInfo": {
            "type": "object",
            "properties": {
              "fullName": {
                "type": "string"
              },
              "email": {
                "type": "string"
              },
              "phone": {
                "type": "string"
              },
              "address": {
                "type": "string"
              },
              "city": {
                "type": "string"
              }
            }
          },
          "payment": {
            "type": "object",
            "properties": {
              "method": {
                "type": "string",
                "enum": [
                  "bank_transfer",
                  "credit_card",
                  "cash"
                ]
              },
              "status": {
                "type": "string",
                "enum": [
                  "pending",
                  "completed",
                  "failed",
                  "refunded"
                ]
              },
              "amount": {
                "type": "number"
              }
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "confirmed",
              "processing",
              "completed",
              "cancelled"
            ],
            "example": "pending"
          },
          "totalAmount": {
            "type": "number",
            "description": "Tổng tiền",
            "example": 100000000
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "description": "Thông báo lỗi"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "description": "Chi tiết lỗi validation"
          }
        }
      },
      "Success": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "Thông báo thành công",
            "example": "Thành công"
          },
          "data": {
            "type": "object",
            "description": "Dữ liệu trả về"
          },
          "count": {
            "type": "number",
            "description": "Số lượng items trong trang hiện tại",
            "example": 10
          },
          "total": {
            "type": "number",
            "description": "Tổng số items",
            "example": 100
          },
          "pagination": {
            "type": "object",
            "properties": {
              "next": {
                "type": "object",
                "properties": {
                  "page": {
                    "type": "number",
                    "example": 2
                  },
                  "limit": {
                    "type": "number",
                    "example": 10
                  }
                }
              },
              "prev": {
                "type": "object",
                "properties": {
                  "page": {
                    "type": "number",
                    "example": 1
                  },
                  "limit": {
                    "type": "number",
                    "example": 10
                  }
                }
              }
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/orders/my-orders": {
      "get": {
        "summary": "Lấy danh sách đơn hàng của user hiện tại",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Số trang"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            },
            "description": "Số lượng đơn hàng mỗi trang"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "confirmed",
                "processing",
                "completed",
                "cancelled"
              ]
            },
            "description": "Lọc theo trạng thái"
          }
        ],
        "responses": {
          "200": {
            "description": "Danh sách đơn hàng",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Success"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Order"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Không có quyền truy cập",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/orders": {
      "post": {
        "summary": "Tạo đơn hàng mới",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "items",
                  "customerInfo",
                  "payment"
                ],
                "properties": {
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "car",
                        "quantity",
                        "selectedColor",
                        "depositAmount"
                      ],
                      "properties": {
                        "car": {
                          "type": "string",
                          "description": "ID của xe",
                          "example": "60d5ecb74b24a1234567890a"
                        },
                        "quantity": {
                          "type": "integer",
                          "minimum": 1,
                          "example": 1
                        },
                        "selectedColor": {
                          "type": "string",
                          "example": "Trắng"
                        },
                        "depositAmount": {
                          "type": "number",
                          "minimum": 0,
                          "example": 50000000
                        }
                      }
                    }
                  },
                  "customerInfo": {
                    "type": "object",
                    "required": [
                      "fullName",
                      "email",
                      "phone",
                      "address",
                      "city"
                    ],
                    "properties": {
                      "fullName": {
                        "type": "string",
                        "example": "Nguyễn Văn A"
                      },
                      "email": {
                        "type": "string",
                        "format": "email",
                        "example": "customer@example.com"
                      },
                      "phone": {
                        "type": "string",
                        "pattern": "^[0-9]{10,11}$",
                        "example": "0123456789"
                      },
                      "address": {
                        "type": "string",
                        "example": "123 ABC Street"
                      },
                      "city": {
                        "type": "string",
                        "example": "Hà Nội"
                      }
                    }
                  },
                  "payment": {
                    "type": "object",
                    "required": [
                      "method"
                    ],
                    "properties": {
                      "method": {
                        "type": "string",
                        "enum": [
                          "bank_transfer",
                          "credit_card",
                          "cash"
                        ],
                        "example": "bank_transfer"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tạo đơn hàng thành công",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Success"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Order"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu không hợp lệ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Không có quyền truy cập",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/orders/{id}": {
      "get": {
        "summary": "Lấy chi tiết đơn hàng",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID của đơn hàng"
          }
        ],
        "responses": {
          "200": {
            "description": "Chi tiết đơn hàng",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Success"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Order"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Không có quyền truy cập đơn hàng này",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy đơn hàng",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}